Unify Text function into Inbox, simplify namespaces, move code
Make each message its own file, remove older separation between Text and Image
messages for now (didn't seem overly useful).
