+++ Api/Core/Messaging.cs
@@ -2,7 +2,6 @@ using System;
 using System.Net.Http;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using NosAyudamos.Events;
 using Polly.Registry;
 
 namespace NosAyudamos
