+++ Api/Domain/MessageReceived.cs
@@ -1,0 +1,10 @@
+﻿namespace NosAyudamos
+{
+    public class MessageReceived : MessageEvent
+    {
+        public MessageReceived(string from, string to, string body) : base(from, to)
+            => Body = body;
+
+        public string Body { get; }
+    }
+}
