+++ Api/Functions/ChatApiIncoming.cs
@@ -6,7 +6,6 @@ using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Azure.WebJobs;
 using Microsoft.Azure.WebJobs.Extensions.Http;
-using NosAyudamos.Events;
 
 namespace NosAyudamos.Functions
 {
