+++ Api/Workflow/DoneeWorkflow.cs
@@ -1,7 +1,6 @@
 using System.Composition;
 using System.Threading.Tasks;
 using Microsoft.Extensions.Logging;
-using NosAyudamos.Events;
 
 namespace NosAyudamos
 {
