+++ Api/Workflow/IWorkflow.cs
@@ -1,5 +1,4 @@
 using System.Threading.Tasks;
-using NosAyudamos.Events;
 
 namespace NosAyudamos
 {
