+++ Api/Workflow/StartupWorkflow.cs
@@ -2,8 +2,8 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.Net.Http;
 using System.Threading.Tasks;
+using Microsoft.Azure.Documents.SystemFunctions;
 using Microsoft.Extensions.Logging;
-using NosAyudamos.Events;
 
 namespace NosAyudamos
 {
@@ -41,20 +41,20 @@ namespace NosAyudamos
             => (this.enviroment, this.languageUnderstanding, this.messaging, this.workflowSelector, this.blobStorage, this.idRecognizer, this.repositoryFactory, this.personRepository, this.httpClient, this.interactiveAction, this.logger)
             = (enviroment, languageUnderstanding, messaging, workflowSelector, blobStorage, idRecognizer, repositoryFactory, personRepository, httpClient, interactiveAction, logger);
 
-        public async Task RunAsync(MessageEvent @event) => await RegisterDoneeAsync((ImageMessageReceived)@event);
+        public async Task RunAsync(MessageEvent @event) => await RegisterDoneeAsync((MessageReceived)@event);
 
-        private async Task RegisterDoneeAsync(ImageMessageReceived e)
+        private async Task RegisterDoneeAsync(MessageReceived e)
         {
 #pragma warning disable CS8634 // The type cannot be used as type parameter in the generic type or method. Nullability of type argument doesn't match 'class' constraint.
             var id = await interactiveAction.ExecuteAsync(
 #pragma warning restore CS8634 // The type cannot be used as type parameter in the generic type or method. Nullability of type argument doesn't match 'class' constraint.
-                () => idRecognizer.RecognizeAsync(e.ImageUri),
+                () => idRecognizer.RecognizeAsync(e.Body),
                 () => messaging.SendTextAsync(e.To, "Dni invalido, intente de nuevo.", e.From),
                 () => messaging.SendTextAsync(e.To, "No pudimos procesar su dni. Nos contactaremos en breve.", e.From));
 
             if (id != null)
             {
-                var image = await httpClient.GetByteArrayAsync(e.ImageUri);
+                var image = await httpClient.GetByteArrayAsync(e.Body);
 
                 await blobStorage.UploadAsync(
                     image, enviroment.GetVariable("AttachmentsContainerName"), $"dni_{id.NationalId}.png");
